remove a if clause

+++ lib/vagrant-zones/driver.rb
@@ -140,34 +140,32 @@ module VagrantPlugins
                      when /host/
                        'h'
                      end
-          if adpatertype.to_s == 'public_network'
-            if opts[:dhcp] && opts[:managed]
-              vnic_name = "vnic#{nic_type}#{config.vm_type}_#{config.partition_id}_#{nic_number}"
-              PTY.spawn("pfexec zlogin -C #{name}") do |zlogin_read, zlogin_write, pid|
-                command = "ip -4 addr show dev #{vnic_name} | head -n -1 | tail -1  | awk '{ print $2 }'  | cut -f1 -d\"/\" \n"
-                zlogin_read.expect(/\n/) { zlogin_write.printf(command) }
-                Timeout.timeout(30) do
-                  loop do
-                    zlogin_read.expect(/\r\n/) { |line| responses.push line }
-                    if responses[-1].to_s.match(/(?:[0-9]{1,3}\.){3}[0-9]{1,3}/)
-                      ip = responses[-1][0].rstrip.gsub(/\e\[\?2004l/, '').lstrip
-                      return nil if ip.empty?
-                      return ip.gsub(/\t/, '') unless ip.empty?
+          if opts[:dhcp] && opts[:managed] && adpatertype.to_s == 'public_network'
+            vnic_name = "vnic#{nic_type}#{config.vm_type}_#{config.partition_id}_#{nic_number}"
+            PTY.spawn("pfexec zlogin -C #{name}") do |zlogin_read, zlogin_write, pid|
+              command = "ip -4 addr show dev #{vnic_name} | head -n -1 | tail -1  | awk '{ print $2 }'  | cut -f1 -d\"/\" \n"
+              zlogin_read.expect(/\n/) { zlogin_write.printf(command) }
+              Timeout.timeout(30) do
+                loop do
+                  zlogin_read.expect(/\r\n/) { |line| responses.push line }
+                  if responses[-1].to_s.match(/(?:[0-9]{1,3}\.){3}[0-9]{1,3}/)
+                    ip = responses[-1][0].rstrip.gsub(/\e\[\?2004l/, '').lstrip
+                    return nil if ip.empty?
+                    return ip.gsub(/\t/, '') unless ip.empty?
 
-                      break
-                    end
-                    errormessage = "==> #{name} ==> Command ==> #{cmd} \nFailed with ==> #{responses[-1]}"
-                    raise errormessage if responses[-1].to_s.match(/Error Code: \b(?!0\b)\d{1,4}\b/)
+                    break
                   end
+                  errormessage = "==> #{name} ==> Command ==> #{cmd} \nFailed with ==> #{responses[-1]}"
+                  raise errormessage if responses[-1].to_s.match(/Error Code: \b(?!0\b)\d{1,4}\b/)
                 end
-                Process.kill('HUP', pid)
               end
-            elsif (opts[:dhcp] == false || opts[:dhcp].nil?) && opts[:managed]
-              ip = opts[:ip].to_s
-              return nil if ip.empty?
-
-              return ip.gsub(/\t/, '')
+              Process.kill('HUP', pid)
             end
+          elsif (opts[:dhcp] == false || opts[:dhcp].nil?) && opts[:managed] && adpatertype.to_s == 'public_network'
+            ip = opts[:ip].to_s
+            return nil if ip.empty?
+
+            return ip.gsub(/\t/, '')
           end
         end
       end
